Fix CORS `allowed_origins` port. (#47)
When using the dashboard -- served on `localhost:8080` -- the browser
will include the `Origin` header and will specify `localhost:8080`, so
this value must be explicitly accepted by the sioserver. Currently the
sioserver is serving on `localhost:8081`, so even though the host is the
same, the port is different and using `8081` results in a dashboard
error.

The problem does not subsist for other Python clients/sensors, since the
websocket connection is not initiated by a browser and doesn't send an
`Origin` header.