Use correct var names when reading the config. (#50)
This PR fixes two issues that prevented the server from running
properly:
* `sioserver.py` is currently looking for `name` and `description` in
the config, even though we settled on `name` and `desc`
* the `sensor_info`s use `sensor_name` and `sensor_desc`, so when
comparing with those, the `sensor_` prefix is added

The logic also seems backward, since the `sensor_info` values take
precedence over the config file, whereas it should be the opposite. I'll
fix this in a separate PR after some additional testing.