Make `emit_readings()` more robust. (#53)
This should fix a somewhat rare issue that I've encountered:

https://github.com/overthesun/simoc-sam/blob/65156f94fa91bcf9ec63ba2d628beda54bd8789e/src/simoc_sam/sioserver.py#L181-L184
This piece of code executes `emit_readings` when the server starts, and
`emit_readings` loops forever in a background task, sending readings to
the clients:

https://github.com/overthesun/simoc-sam/blob/65156f94fa91bcf9ec63ba2d628beda54bd8789e/src/simoc_sam/sioserver.py#L150-L165
This in turn calls `emit_to_subscribers`, that sends the readings to
each subscriber one by one:

https://github.com/overthesun/simoc-sam/blob/65156f94fa91bcf9ec63ba2d628beda54bd8789e/src/simoc_sam/sioserver.py#L113-L116

My understanding is that this background task is executed in a separate
thread, but it's not thread safe. If a new subscriber is added or
removed while `emit_to_subscribers` is executing its loop, the code will
fail because the `SUBSCRIBERS` set changed size during iteration.

It seems that sometimes, the frontend loses the connection. I think this
might happen when the CPU usage is so high that the dashboard is too
slow to reply and something timeouts. This was observed on the macmini
(which is slow), and locally, by reducing the delay at which the
sioserver sent readings to the client to 0.001s (instead of 1s at the
end of `emit_readings`).

If this happens, the `try`/`except` will catch and ignore the error. I
also added an additional check (`if readings`) so that `emit_readings`
doesn't try to read values from a sensor that hasn't produced any value
yet.