Improve tmux layout and message logging. (#54)
This PR improves the tmux layout to look like this:

![image](https://user-images.githubusercontent.com/25624924/233602844-2fd55a62-9027-48fc-9119-4fbb5479a807.png)

The server (on the left) now only prints a message every few broadcasts,
without printing the actual readings. It also prints how many
sensors/clients are connected. On the right there is a panel for the
sensors (during mission-1 there will only be Vernier sensors), one for
the sioclient/csv logger, and a shell (useful for debugging). These 3
panels are now wider so that more data fit in one line.

I also commented out a few more redundant messages.