Avoid using relative cwd in the copy-repo command. (#70)
This PR changes the behavior of the `copy-repo` command so that instead
of copying the cwd, it always copies the root of the repo. Note that
`rsync` behavior changes depending on whether the source has a trailing
`/` or not:
* without trailing `/` it copies the directory with all its content
* with a trailing `/` it only copies the content

Here we want to copy/update the content of the `simoc-sam` dir, hence
the trailing `/`.