Fallback on system `python` if the venv is missing. (#73)
The `sam` alias defined in `bash_aliases` uses the `python` defined in
the venv, whether that's necessary or not. This works fine for most
commands, except `create-venv`, since the venv's `python` is not
available and the command fails.

This PR tries to use the `python` defined in the venv, and if it doesn't
exist it falls back on the system `python`.