Add a systemd unit file to run sensors as services (#112)
Instead of trying to launch `tmux` as a service, I created a unit file
template that can be used to launch the individual sensor scripts as
services, that can be enabled to automatically start on boot. The logs
can be inspected using e.g. `journalctl -u sensor-runner@scd30.service
-f`, so technically a `tmux` script that launches `journalctl` could be
created to inspect multiple logs at once. This PR doesn't include this,
and also doesn't include automation to (un)symlink this file. This can
be added in a separate PR.