Test with multiple Python versions (#129)
This PR updates the GitHub workflow to test with different Python
versions.

Currently we are using:
* `3.9` on the RPi 0s
* `3.11` on the RPi 4s

We also have a server that still has `3.6`, but I'm not sure we want to
limit ourselves to such an old version, especially if we manage to
update the server (or switch to a different server). Versions `>3.11`
are added to make sure the code runs fine on more modern Python
versions.