Add a command to install the touchscreen. (#130)
This `simoc-sam.py` installs the [GeeekPi 3.5" touchscreen for the
Raspberry Pi 4/5](https://www.amazon.com/dp/B0D7VDWBBC).

In order to do that it will:
* clone https://github.com/goodtft/LCD-show.git
* make the whole repo executable (a bit sketchy, but the instructions
says to do it)
* run the
[`MHS35-show`](https://github.com/goodtft/LCD-show/blob/master/MHS35-show)
script (that does more sketchy things to install the screen)
* schedule the repo to be deleted on boot (since the scripts reboots the
Pi before `simoc-sam.py` does it)

This is currently untested, but will test it and report back shortly.