Set 960x640 as default res for the 3.5" screen (#133)
This is a follow-up of:
* #130 

Instead of using the default 480x320 resolution (which is quite
cramped), it doubles it to 960x320. Even though the screen only has
480x320 pixels, the image is then downscaled automatically but looks
fine, and the screen can still fit 4 times more things.

We could also add a command to change the resolution that calls e.g.
`fbset -xres 960 -yres 640`, but so far it doesn't seem particularly
useful.