Generate the list of CORS allowed origins dynamically (#209)
This PR updates the code in `siobridge.py` to automatically find all IPs
and hostname that point to the local host and add them to the
`allowed_origins`. This will allow anyone in the same private networks
of the host running the `siobridge` and SIMOC Web to access the web
interface regardless of the network they are in.