Preserve HOME envvar while running as sudo (#229)
The `HOME` env var is used by `Path.home()` to find files such as the
user config introduced in #214. Some `simoc-sam.py` commands are
decorated with `@needs_root`, that run the command with `sudo`. Running
commands with `sudo` changes the current user, which changes the value
of `HOME` and affects the result of `Path.home()`. This PR fixes
`@needs_root` to preserve the original `HOME`.