Use the hostname to set the config.location (#242)
Since the `hostname` can be set from the RPi imager, we can use it to
set the `config.location` dynamically without having to SSH into the RPi
and editing the user `config.py` manually.

This PR implements this functionality, stripping the trailing numbers so
that multiple sensor arrays in the same hab will have the same
`location` (assuming they are named `locationN`).